requirement.txt
alembic==1.7.6
aniso8601==9.0.1
attrs==21.4.0
Authlib==0.15.5
Babel==2.9.1
bcrypt==3.2.0
blinker==1.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
cryptography==36.0.1
dnspython==2.2.0
email-validator==1.1.3
Flask==2.0.3
Flask-BabelEx==0.9.4
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.10
Flask-DebugToolbar==0.11.0
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==3.1.0
Flask-OAuth==0.12
Flask-Principal==0.4.0
Flask-RESTful==0.3.9
flask-restplus==0.13.0
Flask-Security==3.0.0
Flask-Social==1.6.2
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.0
greenlet==1.1.2
httplib2==0.20.4
idna==3.3
is-disposable-email==1.0.0
itsdangerous==2.1.0
Jinja2==3.0.3
jsonschema==4.4.0
Mako==1.1.6
MarkupSafe==2.1.0
oauth2==1.9.0.post1
oauthlib==3.2.0
passlib==1.7.4
pycparser==2.21
pyparsing==3.0.7
pyrsistent==0.18.1
pytz==2021.3
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
speaklater==1.3
SQLAlchemy==1.4.31
urllib3==1.26.8
Werkzeug==2.0.3
WTForms==3.0.1
wtforms-validators==1.0.0